REST API/working/Entso-E API/requirements.txt
entsoe_py==0.4.2
pandas==1.3.5
python-dotenv==0.21.0
